Moved the python generator and serialization support into separate packages
